ProjetoFinal/Worker/requirements.txt
boto3==1.34.0
python-dotenv==1.0.0
Flask==3.0.0
